requirements.txt
neo4j==5.22.0
python-dotenv==1.0.1
mcp>=1.23.0
requests>=2.31.0
flask>=3.0.0
flask-cors>=4.0.0